requirements.txt
fastapi
requests
uvicorn
browser-use
python-dotenv
boto3 